requirements.txt
# Web scraping and parsing
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
tabula-py==2.9.0

# Text processing and chunking
markdown==3.5.1
html2text==2020.1.16
tiktoken==0.5.1

# Vector embeddings and search
openai==1.3.5
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Data handling
pandas==2.1.3
numpy==1.26.2

# Web framework
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
